(feat) moved some files and created exe using auto-py-to-exe
